requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
django-ninja==0.22.2
Markdown==3.5.1
pydantic==1.10.13
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
